Behavior-preserving refactor of `Value` into a module. (#509)
