+++ partiql-value/src/lib.rs
@@ -10,22 +10,26 @@ use std::fmt::{Debug, Display, Formatter};
 use std::hash::Hash;
 
 use std::iter::Once;
-use std::{ops, vec};
 
 use rust_decimal::prelude::FromPrimitive;
-use rust_decimal::{Decimal as RustDecimal, Decimal};
+use rust_decimal::Decimal as RustDecimal;
 
 mod bag;
+pub mod comparison;
 mod datetime;
 mod list;
 mod pretty;
 mod tuple;
+mod util;
+mod value;
 
 pub use bag::*;
+pub use comparison::*;
 pub use datetime::*;
 pub use list::*;
 pub use pretty::*;
 pub use tuple::*;
+pub use value::*;
 
 use partiql_common::pretty::ToPretty;
 #[cfg(feature = "serde")]
@@ -36,702 +40,6 @@ use serde::{Deserialize, Serialize};
 pub enum BindingsName<'s> {
     CaseSensitive(Cow<'s, str>),
     CaseInsensitive(Cow<'s, str>),
-}
-
-// TODO these are all quite simplified for PoC/demonstration
-// TODO have an optional-like wrapper for null/missing instead of inlined here?
-#[derive(Hash, PartialEq, Eq, Clone)]
-#[allow(dead_code)] // TODO remove once out of PoC
-#[derive(Default)]
-#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-pub enum Value {
-    Null,
-    #[default]
-    Missing,
-    Boolean(bool),
-    Integer(i64),
-    Real(OrderedFloat<f64>),
-    Decimal(Box<RustDecimal>),
-    String(Box<String>),
-    Blob(Box<Vec<u8>>),
-    DateTime(Box<DateTime>),
-    List(Box<List>),
-    Bag(Box<Bag>),
-    Tuple(Box<Tuple>),
-    // TODO: add other supported PartiQL values -- sexp
-}
-
-impl Display for Value {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        match self.to_pretty_string(f.width().unwrap_or(80)) {
-            Ok(pretty) => f.write_str(&pretty),
-            Err(_) => f.write_str("<internal value error occurred>"),
-        }
-    }
-}
-
-impl ops::Add for &Value {
-    type Output = Value;
-
-    fn add(self, rhs: Self) -> Self::Output {
-        match (&self, &rhs) {
-            // TODO: edge cases dealing with overflow
-            (Value::Missing, _) => Value::Missing,
-            (_, Value::Missing) => Value::Missing,
-            (Value::Null, _) => Value::Null,
-            (_, Value::Null) => Value::Null,
-            (Value::Integer(l), Value::Integer(r)) => Value::Integer(l + r),
-            (Value::Real(l), Value::Real(r)) => Value::Real(*l + *r),
-            (Value::Decimal(l), Value::Decimal(r)) => {
-                Value::Decimal(Box::new(l.as_ref() + r.as_ref()))
-            }
-            (Value::Integer(_), Value::Real(_)) => &coerce_int_to_real(self) + rhs,
-            (Value::Integer(_), Value::Decimal(_)) => &coerce_int_or_real_to_decimal(self) + rhs,
-            (Value::Real(_), Value::Decimal(_)) => &coerce_int_or_real_to_decimal(self) + rhs,
-            (Value::Real(_), Value::Integer(_)) => self + &coerce_int_to_real(rhs),
-            (Value::Decimal(_), Value::Integer(_)) => self + &coerce_int_or_real_to_decimal(rhs),
-            (Value::Decimal(_), Value::Real(_)) => self + &coerce_int_or_real_to_decimal(rhs),
-            _ => Value::Missing, // data type mismatch => Missing
-        }
-    }
-}
-
-impl ops::AddAssign<&Value> for Value {
-    fn add_assign(&mut self, rhs: &Value) {
-        match (self, &rhs) {
-            // TODO: edge cases dealing with overflow
-            (Value::Missing, _) => {}
-            (this, Value::Missing) => *this = Value::Missing,
-            (Value::Null, _) => {}
-            (this, Value::Null) => *this = Value::Null,
-
-            (Value::Integer(l), Value::Integer(r)) => l.add_assign(r),
-
-            (Value::Real(l), Value::Real(r)) => l.add_assign(r),
-            (Value::Real(l), Value::Integer(i)) => l.add_assign(*i as f64),
-
-            (Value::Decimal(l), Value::Decimal(r)) => l.add_assign(r.as_ref()),
-            (Value::Decimal(l), Value::Integer(i)) => l.add_assign(rust_decimal::Decimal::from(*i)),
-            (Value::Decimal(l), Value::Real(r)) => match coerce_f64_to_decimal(r) {
-                Some(d) => l.add_assign(d),
-                None => todo!(),
-            },
-
-            (this, Value::Real(r)) => {
-                *this = match &this {
-                    Value::Integer(l) => Value::from((*l as f64) + r.0),
-                    _ => Value::Missing,
-                };
-            }
-            (this, Value::Decimal(r)) => {
-                *this = match &this {
-                    Value::Integer(l) => {
-                        Value::Decimal(Box::new(rust_decimal::Decimal::from(*l) + r.as_ref()))
-                    }
-                    Value::Real(l) => match coerce_f64_to_decimal(&l.0) {
-                        None => Value::Missing,
-                        Some(d) => Value::Decimal(Box::new(d + r.as_ref())),
-                    },
-                    _ => Value::Missing,
-                };
-            }
-            (this, _) => *this = Value::Missing, // data type mismatch => Missing
-        }
-    }
-}
-
-impl ops::Sub for &Value {
-    type Output = Value;
-
-    fn sub(self, rhs: Self) -> Self::Output {
-        match (&self, &rhs) {
-            // TODO: edge cases dealing with overflow
-            (Value::Missing, _) => Value::Missing,
-            (_, Value::Missing) => Value::Missing,
-            (Value::Null, _) => Value::Null,
-            (_, Value::Null) => Value::Null,
-            (Value::Integer(l), Value::Integer(r)) => Value::Integer(l - r),
-            (Value::Real(l), Value::Real(r)) => Value::Real(*l - *r),
-            (Value::Decimal(l), Value::Decimal(r)) => {
-                Value::Decimal(Box::new(l.as_ref() - r.as_ref()))
-            }
-            (Value::Integer(_), Value::Real(_)) => &coerce_int_to_real(self) - rhs,
-            (Value::Integer(_), Value::Decimal(_)) => &coerce_int_or_real_to_decimal(self) - rhs,
-            (Value::Real(_), Value::Decimal(_)) => &coerce_int_or_real_to_decimal(self) - rhs,
-            (Value::Real(_), Value::Integer(_)) => self - &coerce_int_to_real(rhs),
-            (Value::Decimal(_), Value::Integer(_)) => self - &coerce_int_or_real_to_decimal(rhs),
-            (Value::Decimal(_), Value::Real(_)) => self - &coerce_int_or_real_to_decimal(rhs),
-            _ => Value::Missing, // data type mismatch => Missing
-        }
-    }
-}
-
-impl ops::Mul for &Value {
-    type Output = Value;
-
-    fn mul(self, rhs: Self) -> Self::Output {
-        match (&self, &rhs) {
-            // TODO: edge cases dealing with overflow
-            (Value::Missing, _) => Value::Missing,
-            (_, Value::Missing) => Value::Missing,
-            (Value::Null, _) => Value::Null,
-            (_, Value::Null) => Value::Null,
-            (Value::Integer(l), Value::Integer(r)) => Value::Integer(l * r),
-            (Value::Real(l), Value::Real(r)) => Value::Real(*l * *r),
-            (Value::Decimal(l), Value::Decimal(r)) => {
-                Value::Decimal(Box::new(l.as_ref() * r.as_ref()))
-            }
-            (Value::Integer(_), Value::Real(_)) => &coerce_int_to_real(self) * rhs,
-            (Value::Integer(_), Value::Decimal(_)) => &coerce_int_or_real_to_decimal(self) * rhs,
-            (Value::Real(_), Value::Decimal(_)) => &coerce_int_or_real_to_decimal(self) * rhs,
-            (Value::Real(_), Value::Integer(_)) => self * &coerce_int_to_real(rhs),
-            (Value::Decimal(_), Value::Integer(_)) => self * &coerce_int_or_real_to_decimal(rhs),
-            (Value::Decimal(_), Value::Real(_)) => self * &coerce_int_or_real_to_decimal(rhs),
-            _ => Value::Missing, // data type mismatch => Missing
-        }
-    }
-}
-
-impl ops::Div for &Value {
-    type Output = Value;
-
-    fn div(self, rhs: Self) -> Self::Output {
-        match (&self, &rhs) {
-            // TODO: edge cases dealing with division by 0
-            (Value::Missing, _) => Value::Missing,
-            (_, Value::Missing) => Value::Missing,
-            (Value::Null, _) => Value::Null,
-            (_, Value::Null) => Value::Null,
-            (Value::Integer(l), Value::Integer(r)) => Value::Integer(l / r),
-            (Value::Real(l), Value::Real(r)) => Value::Real(*l / *r),
-            (Value::Decimal(l), Value::Decimal(r)) => {
-                Value::Decimal(Box::new(l.as_ref() / r.as_ref()))
-            }
-            (Value::Integer(_), Value::Real(_)) => &coerce_int_to_real(self) / rhs,
-            (Value::Integer(_), Value::Decimal(_)) => &coerce_int_or_real_to_decimal(self) / rhs,
-            (Value::Real(_), Value::Decimal(_)) => &coerce_int_or_real_to_decimal(self) / rhs,
-            (Value::Real(_), Value::Integer(_)) => self / &coerce_int_to_real(rhs),
-            (Value::Decimal(_), Value::Integer(_)) => self / &coerce_int_or_real_to_decimal(rhs),
-            (Value::Decimal(_), Value::Real(_)) => self / &coerce_int_or_real_to_decimal(rhs),
-            _ => Value::Missing, // data type mismatch => Missing
-        }
-    }
-}
-
-impl ops::Rem for &Value {
-    type Output = Value;
-
-    fn rem(self, rhs: Self) -> Self::Output {
-        match (&self, &rhs) {
-            // TODO: edge cases dealing with division by 0
-            (Value::Missing, _) => Value::Missing,
-            (_, Value::Missing) => Value::Missing,
-            (Value::Null, _) => Value::Null,
-            (_, Value::Null) => Value::Null,
-            (Value::Integer(l), Value::Integer(r)) => Value::Integer(l % r),
-            (Value::Real(l), Value::Real(r)) => Value::Real(*l % *r),
-            (Value::Decimal(l), Value::Decimal(r)) => {
-                Value::Decimal(Box::new(l.as_ref() % r.as_ref()))
-            }
-            (Value::Integer(_), Value::Real(_)) => &coerce_int_to_real(self) % rhs,
-            (Value::Integer(_), Value::Decimal(_)) => &coerce_int_or_real_to_decimal(self) % rhs,
-            (Value::Real(_), Value::Decimal(_)) => &coerce_int_or_real_to_decimal(self) % rhs,
-            (Value::Real(_), Value::Integer(_)) => self % &coerce_int_to_real(rhs),
-            (Value::Decimal(_), Value::Integer(_)) => self % &coerce_int_or_real_to_decimal(rhs),
-            (Value::Decimal(_), Value::Real(_)) => self % &coerce_int_or_real_to_decimal(rhs),
-            _ => Value::Missing, // data type mismatch => Missing
-        }
-    }
-}
-
-pub trait UnaryPlus {
-    type Output;
-
-    fn positive(self) -> Self::Output;
-}
-
-impl UnaryPlus for Value {
-    type Output = Self;
-    fn positive(self) -> Self::Output {
-        match self {
-            Value::Null => Value::Null,
-            Value::Missing => Value::Missing,
-            Value::Integer(_) | Value::Real(_) | Value::Decimal(_) => self,
-            _ => Value::Missing, // data type mismatch => Missing
-        }
-    }
-}
-
-impl ops::Neg for &Value {
-    type Output = Value;
-
-    fn neg(self) -> Self::Output {
-        match self {
-            // TODO: handle overflow for negation
-            Value::Null => Value::Null,
-            Value::Missing => Value::Missing,
-            Value::Integer(i) => Value::from(-i),
-            Value::Real(f) => Value::Real(-f),
-            Value::Decimal(d) => Value::from(-d.as_ref()),
-            _ => Value::Missing, // data type mismatch => Missing
-        }
-    }
-}
-
-impl ops::Neg for Value {
-    type Output = Value;
-
-    fn neg(self) -> Self::Output {
-        match self {
-            // TODO: handle overflow for negation
-            Value::Null => self,
-            Value::Missing => self,
-            Value::Integer(i) => Value::from(-i),
-            Value::Real(f) => Value::Real(-f),
-            Value::Decimal(d) => Value::from(-d.as_ref()),
-            _ => Value::Missing, // data type mismatch => Missing
-        }
-    }
-}
-
-pub trait BinaryAnd {
-    type Output;
-
-    fn and(&self, rhs: &Self) -> Self::Output;
-}
-
-impl BinaryAnd for Value {
-    type Output = Self;
-    fn and(&self, rhs: &Self) -> Self::Output {
-        match (self, rhs) {
-            (Value::Boolean(l), Value::Boolean(r)) => Value::from(*l && *r),
-            (Value::Null | Value::Missing, Value::Boolean(false))
-            | (Value::Boolean(false), Value::Null | Value::Missing) => Value::from(false),
-            _ => {
-                if matches!(self, Value::Missing | Value::Null | Value::Boolean(true))
-                    && matches!(rhs, Value::Missing | Value::Null | Value::Boolean(true))
-                {
-                    Value::Null
-                } else {
-                    Value::Missing
-                }
-            }
-        }
-    }
-}
-
-pub trait BinaryOr {
-    type Output;
-
-    fn or(&self, rhs: &Self) -> Self::Output;
-}
-
-impl BinaryOr for Value {
-    type Output = Self;
-    fn or(&self, rhs: &Self) -> Self::Output {
-        match (self, rhs) {
-            (Value::Boolean(l), Value::Boolean(r)) => Value::from(*l || *r),
-            (Value::Null | Value::Missing, Value::Boolean(true))
-            | (Value::Boolean(true), Value::Null | Value::Missing) => Value::from(true),
-            _ => {
-                if matches!(self, Value::Missing | Value::Null | Value::Boolean(false))
-                    && matches!(rhs, Value::Missing | Value::Null | Value::Boolean(false))
-                {
-                    Value::Null
-                } else {
-                    Value::Missing
-                }
-            }
-        }
-    }
-}
-
-impl ops::Not for &Value {
-    type Output = Value;
-
-    fn not(self) -> Self::Output {
-        match self {
-            Value::Boolean(b) => Value::from(!b),
-            Value::Null | Value::Missing => Value::Null,
-            _ => Value::Missing, // data type mismatch => Missing
-        }
-    }
-}
-
-impl ops::Not for Value {
-    type Output = Self;
-
-    fn not(self) -> Self::Output {
-        match self {
-            Value::Boolean(b) => Value::from(!b),
-            Value::Null | Value::Missing => Value::Null,
-            _ => Value::Missing, // data type mismatch => Missing
-        }
-    }
-}
-
-pub trait Comparable {
-    fn is_comparable_to(&self, rhs: &Self) -> bool;
-}
-
-impl Comparable for Value {
-    /// Returns true if and only if `self` is comparable to `rhs`
-    fn is_comparable_to(&self, rhs: &Self) -> bool {
-        match (self, rhs) {
-            // Null/Missing compare to anything
-            (Value::Missing | Value::Null, _)
-            | (_, Value::Missing | Value::Null)
-            // Everything compares to its own type
-            | (Value::Boolean(_), Value::Boolean(_))
-            | (Value::String(_), Value::String(_))
-            | (Value::Blob(_), Value::Blob(_))
-            | (Value::List(_), Value::List(_))
-            | (Value::Bag(_), Value::Bag(_))
-            | (Value::Tuple(_), Value::Tuple(_))
-            // Numerics compare to each other
-            | (
-                Value::Integer(_) | Value::Real(_) | Value::Decimal(_),
-                Value::Integer(_) | Value::Real(_) | Value::Decimal(_),
-            )=> true,
-            (_, _) => false,
-        }
-    }
-}
-
-// `Value` `eq` and `neq` with Missing and Null propagation
-pub trait NullableEq {
-    type Output;
-    fn eq(&self, rhs: &Self) -> Self::Output;
-    fn neq(&self, rhs: &Self) -> Self::Output;
-}
-
-// `Value` comparison with Missing and Null propagation
-pub trait NullableOrd {
-    type Output;
-
-    fn lt(&self, rhs: &Self) -> Self::Output;
-    fn gt(&self, rhs: &Self) -> Self::Output;
-    fn lteq(&self, rhs: &Self) -> Self::Output;
-    fn gteq(&self, rhs: &Self) -> Self::Output;
-}
-
-/// A wrapper on [`T`] that specifies if missing and null values should be equal.
-#[derive(Eq, PartialEq)]
-pub struct EqualityValue<'a, const NULLS_EQUAL: bool, T>(pub &'a T);
-
-impl<const GROUP_NULLS: bool> NullableEq for EqualityValue<'_, GROUP_NULLS, Value> {
-    type Output = Value;
-
-    fn eq(&self, rhs: &Self) -> Self::Output {
-        if GROUP_NULLS {
-            if let (Value::Missing | Value::Null, Value::Missing | Value::Null) = (self.0, rhs.0) {
-                return Value::Boolean(true);
-            }
-        } else if matches!(self.0, Value::Missing) || matches!(rhs.0, Value::Missing) {
-            return Value::Missing;
-        } else if matches!(self.0, Value::Null) || matches!(rhs.0, Value::Null) {
-            return Value::Null;
-        }
-
-        match (self.0, rhs.0) {
-            (Value::Integer(_), Value::Real(_)) => {
-                Value::from(&coerce_int_to_real(self.0) == rhs.0)
-            }
-            (Value::Integer(_), Value::Decimal(_)) => {
-                Value::from(&coerce_int_or_real_to_decimal(self.0) == rhs.0)
-            }
-            (Value::Real(_), Value::Decimal(_)) => {
-                Value::from(&coerce_int_or_real_to_decimal(self.0) == rhs.0)
-            }
-            (Value::Real(_), Value::Integer(_)) => {
-                Value::from(self.0 == &coerce_int_to_real(rhs.0))
-            }
-            (Value::Decimal(_), Value::Integer(_)) => {
-                Value::from(self.0 == &coerce_int_or_real_to_decimal(rhs.0))
-            }
-            (Value::Decimal(_), Value::Real(_)) => {
-                Value::from(self.0 == &coerce_int_or_real_to_decimal(rhs.0))
-            }
-            (_, _) => Value::from(self.0 == rhs.0),
-        }
-    }
-
-    fn neq(&self, rhs: &Self) -> Self::Output {
-        let eq_result = NullableEq::eq(self, rhs);
-        match eq_result {
-            Value::Boolean(_) | Value::Null => !eq_result,
-            _ => Value::Missing,
-        }
-    }
-}
-
-impl NullableOrd for Value {
-    type Output = Self;
-
-    fn lt(&self, rhs: &Self) -> Self::Output {
-        match (self, rhs) {
-            (Value::Missing, _) => Value::Missing,
-            (_, Value::Missing) => Value::Missing,
-            (Value::Null, _) => Value::Null,
-            (_, Value::Null) => Value::Null,
-            (_, _) => {
-                if self.is_comparable_to(rhs) {
-                    Value::from(self < rhs)
-                } else {
-                    Value::Missing
-                }
-            }
-        }
-    }
-
-    fn gt(&self, rhs: &Self) -> Self::Output {
-        match (self, rhs) {
-            (Value::Missing, _) => Value::Missing,
-            (_, Value::Missing) => Value::Missing,
-            (Value::Null, _) => Value::Null,
-            (_, Value::Null) => Value::Null,
-            (_, _) => {
-                if self.is_comparable_to(rhs) {
-                    Value::from(self > rhs)
-                } else {
-                    Value::Missing
-                }
-            }
-        }
-    }
-
-    fn lteq(&self, rhs: &Self) -> Self::Output {
-        match (self, rhs) {
-            (Value::Missing, _) => Value::Missing,
-            (_, Value::Missing) => Value::Missing,
-            (Value::Null, _) => Value::Null,
-            (_, Value::Null) => Value::Null,
-            (_, _) => {
-                if self.is_comparable_to(rhs) {
-                    Value::from(self <= rhs)
-                } else {
-                    Value::Missing
-                }
-            }
-        }
-    }
-
-    fn gteq(&self, rhs: &Self) -> Self::Output {
-        match (self, rhs) {
-            (Value::Missing, _) => Value::Missing,
-            (_, Value::Missing) => Value::Missing,
-            (Value::Null, _) => Value::Null,
-            (_, Value::Null) => Value::Null,
-            (_, _) => {
-                if self.is_comparable_to(rhs) {
-                    Value::from(self >= rhs)
-                } else {
-                    Value::Missing
-                }
-            }
-        }
-    }
-}
-
-fn coerce_f64_to_decimal(real_value: &f64) -> Option<Decimal> {
-    if !real_value.is_finite() {
-        None
-    } else {
-        Decimal::from_f64(*real_value)
-    }
-}
-
-fn coerce_int_or_real_to_decimal(value: &Value) -> Value {
-    match value {
-        Value::Integer(int_value) => Value::from(rust_decimal::Decimal::from(*int_value)),
-        Value::Real(real_value) => {
-            if !real_value.is_finite() {
-                Value::Missing
-            } else {
-                match Decimal::from_f64(real_value.0) {
-                    Some(d_from_r) => Value::from(d_from_r),
-                    None => Value::Missing, // TODO: decide on behavior when float cannot be coerced to Decimal
-                }
-            }
-        }
-        _ => todo!("Unsupported coercion to Decimal"),
-    }
-}
-
-fn coerce_int_to_real(value: &Value) -> Value {
-    match value {
-        Value::Integer(int_value) => Value::Real(OrderedFloat(*int_value as f64)),
-        _ => todo!("Unsupported coercion to Real"),
-    }
-}
-
-impl Value {
-    #[inline]
-    #[must_use]
-    pub fn is_tuple(&self) -> bool {
-        matches!(self, Value::Tuple(_))
-    }
-
-    #[inline]
-    #[must_use]
-    pub fn is_list(&self) -> bool {
-        matches!(self, Value::List(_))
-    }
-
-    #[inline]
-    #[must_use]
-    pub fn is_bag(&self) -> bool {
-        matches!(self, Value::Bag(_))
-    }
-
-    #[inline]
-    #[must_use]
-    pub fn is_sequence(&self) -> bool {
-        self.is_bag() || self.is_list()
-    }
-
-    #[inline]
-    /// Returns true if and only if Value is an integer, real, or decimal
-    #[must_use]
-    pub fn is_number(&self) -> bool {
-        matches!(self, Value::Integer(_) | Value::Real(_) | Value::Decimal(_))
-    }
-    #[inline]
-    /// Returns true if and only if Value is null or missing
-    #[must_use]
-    pub fn is_absent(&self) -> bool {
-        matches!(self, Value::Missing | Value::Null)
-    }
-
-    #[inline]
-    /// Returns true if Value is neither null nor missing
-    #[must_use]
-    pub fn is_present(&self) -> bool {
-        !self.is_absent()
-    }
-
-    #[inline]
-    #[must_use]
-    pub fn is_ordered(&self) -> bool {
-        self.is_list()
-    }
-
-    #[inline]
-    #[must_use]
-    pub fn coerce_into_tuple(self) -> Tuple {
-        match self {
-            Value::Tuple(t) => *t,
-            _ => self
-                .into_bindings()
-                .map(|(k, v)| (k.unwrap_or_else(|| "_1".to_string()), v))
-                .collect(),
-        }
-    }
-
-    #[inline]
-    #[must_use]
-    pub fn coerce_to_tuple(&self) -> Tuple {
-        match self {
-            Value::Tuple(t) => t.as_ref().clone(),
-            _ => {
-                let fresh = "_1".to_string();
-                self.as_bindings()
-                    .map(|(k, v)| (k.unwrap_or(&fresh), v.clone()))
-                    .collect()
-            }
-        }
-    }
-
-    #[inline]
-    #[must_use]
-    pub fn as_tuple_ref(&self) -> Cow<'_, Tuple> {
-        if let Value::Tuple(t) = self {
-            Cow::Borrowed(t)
-        } else {
-            Cow::Owned(self.coerce_to_tuple())
-        }
-    }
-
-    #[inline]
-    #[must_use]
-    pub fn as_bindings(&self) -> BindingIter<'_> {
-        match self {
-            Value::Tuple(t) => BindingIter::Tuple(t.pairs()),
-            Value::Missing => BindingIter::Empty,
-            _ => BindingIter::Single(std::iter::once(self)),
-        }
-    }
-
-    #[inline]
-    #[must_use]
-    pub fn into_bindings(self) -> BindingIntoIter {
-        match self {
-            Value::Tuple(t) => BindingIntoIter::Tuple(t.into_pairs()),
-            Value::Missing => BindingIntoIter::Empty,
-            _ => BindingIntoIter::Single(std::iter::once(self)),
-        }
-    }
-
-    #[inline]
-    #[must_use]
-    pub fn coerce_into_bag(self) -> Bag {
-        if let Value::Bag(b) = self {
-            *b
-        } else {
-            Bag::from(vec![self])
-        }
-    }
-
-    #[inline]
-    #[must_use]
-    pub fn as_bag_ref(&self) -> Cow<'_, Bag> {
-        if let Value::Bag(b) = self {
-            Cow::Borrowed(b)
-        } else {
-            Cow::Owned(self.clone().coerce_into_bag())
-        }
-    }
-
-    #[inline]
-    #[must_use]
-    pub fn coerce_into_list(self) -> List {
-        if let Value::List(b) = self {
-            *b
-        } else {
-            List::from(vec![self])
-        }
-    }
-
-    #[inline]
-    #[must_use]
-    pub fn as_list_ref(&self) -> Cow<'_, List> {
-        if let Value::List(l) = self {
-            Cow::Borrowed(l)
-        } else {
-            Cow::Owned(self.clone().coerce_into_list())
-        }
-    }
-
-    #[inline]
-    #[must_use]
-    pub fn iter(&self) -> ValueIter<'_> {
-        match self {
-            Value::Null | Value::Missing => ValueIter::Single(None),
-            Value::List(list) => ValueIter::List(list.iter()),
-            Value::Bag(bag) => ValueIter::Bag(bag.iter()),
-            other => ValueIter::Single(Some(other)),
-        }
-    }
-
-    #[inline]
-    #[must_use]
-    pub fn sequence_iter(&self) -> Option<ValueIter<'_>> {
-        if self.is_sequence() {
-            Some(self.iter())
-        } else {
-            None
-        }
-    }
 }
 
 #[derive(Debug, Clone)]
@@ -791,80 +99,15 @@ impl Iterator for BindingIntoIter {
         }
     }
 }
-
-#[derive(Debug, Clone)]
-pub enum ValueIter<'a> {
-    List(ListIter<'a>),
-    Bag(BagIter<'a>),
-    Single(Option<&'a Value>),
-}
-
-impl<'a> Iterator for ValueIter<'a> {
-    type Item = &'a Value;
-
-    #[inline]
-    fn next(&mut self) -> Option<Self::Item> {
-        match self {
-            ValueIter::List(list) => list.next(),
-            ValueIter::Bag(bag) => bag.next(),
-            ValueIter::Single(v) => v.take(),
-        }
-    }
-
-    #[inline]
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        match self {
-            ValueIter::List(list) => list.size_hint(),
-            ValueIter::Bag(bag) => bag.size_hint(),
-            ValueIter::Single(_) => (1, Some(1)),
+impl Display for Value {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        match self.to_pretty_string(f.width().unwrap_or(80)) {
+            Ok(pretty) => f.write_str(&pretty),
+            Err(_) => f.write_str("<internal value error occurred>"),
         }
     }
 }
 
-impl IntoIterator for Value {
-    type Item = Value;
-    type IntoIter = ValueIntoIterator;
-
-    #[inline]
-    fn into_iter(self) -> ValueIntoIterator {
-        match self {
-            Value::List(list) => ValueIntoIterator::List(list.into_iter()),
-            Value::Bag(bag) => ValueIntoIterator::Bag(bag.into_iter()),
-            other => ValueIntoIterator::Single(Some(other)),
-        }
-    }
-}
-
-pub enum ValueIntoIterator {
-    List(ListIntoIterator),
-    Bag(BagIntoIterator),
-    Single(Option<Value>),
-}
-
-impl Iterator for ValueIntoIterator {
-    type Item = Value;
-
-    #[inline]
-    fn next(&mut self) -> Option<Self::Item> {
-        match self {
-            ValueIntoIterator::List(list) => list.next(),
-            ValueIntoIterator::Bag(bag) => bag.next(),
-            ValueIntoIterator::Single(v) => v.take(),
-        }
-    }
-
-    #[inline]
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        match self {
-            ValueIntoIterator::List(list) => list.size_hint(),
-            ValueIntoIterator::Bag(bag) => bag.size_hint(),
-            ValueIntoIterator::Single(_) => (1, Some(1)),
-        }
-    }
-}
-
-// TODO make debug emit proper PartiQL notation
-// TODO perhaps this should be display as well?
 impl Debug for Value {
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
         match self {
@@ -1203,6 +446,7 @@ impl From<Bag> for Value {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::comparison::{EqualityValue, NullableEq, NullableOrd};
     use rust_decimal_macros::dec;
     use std::borrow::Cow;
     use std::cell::RefCell;
@@ -1445,25 +689,25 @@ mod tests {
         assert_eq!(Value::Integer(2), &Value::Integer(1) * &Value::Integer(2));
         assert_eq!(Value::from(3.75), &Value::from(1.5) * &Value::from(2.5));
         assert_eq!(
-            Value::from(Decimal::new(2, 0)),
+            Value::from(RustDecimal::new(2, 0)),
             &Value::Decimal(Box::new(dec!(1))) * &Value::from(dec!(2))
         );
         assert_eq!(Value::from(2.5), &Value::Integer(1) * &Value::from(2.5));
         assert_eq!(Value::from(2.), &Value::from(1.) * &Value::from(2.));
         assert_eq!(
-            Value::from(Decimal::new(2, 0)),
+            Value::from(RustDecimal::new(2, 0)),
             &Value::Integer(1) * &Value::Decimal(Box::new(dec!(2)))
         );
         assert_eq!(
-            Value::from(Decimal::new(2, 0)),
+            Value::from(RustDecimal::new(2, 0)),
             &Value::Decimal(Box::new(dec!(1))) * &Value::Integer(2)
         );
         assert_eq!(
-            Value::from(Decimal::new(2, 0)),
+            Value::from(RustDecimal::new(2, 0)),
             &Value::from(1.) * &Value::Decimal(Box::new(dec!(2)))
         );
         assert_eq!(
-            Value::from(Decimal::new(2, 0)),
+            Value::from(RustDecimal::new(2, 0)),
             &Value::Decimal(Box::new(dec!(1))) * &Value::from(2.)
         );
